+++ app/src/main/java/com/luffy/mulmedia/opengl/TriangleDrawer.java
@@ -109,12 +109,12 @@ public class TriangleDrawer implements IDrawer {
     }
 
     @Override
-    public void translate(int translateX, int translateY) {
+    public void translate(float translateX, float translateY) {
 
     }
 
     @Override
-    public void scale(int scaleX, int scaleY) {
+    public void scale(float scaleX, float scaleY) {
 
     }
